Fixed the FC command problem
The byte arrays we were writing to disk where the size of 65536*512. The first few char were actually letters but the rest where just space. I changed it so that I only writes the chars. There is still a know bug where the last word when doing a write is lost.